- added cells - handled correct user inputs
