+++ webapp/wordle/src/components/Cells/Cells.css
@@ -1,0 +1,14 @@
+.cell-container {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  
+}
+
+.cells {
+  height: 55px;
+  width: 55px;
+  border: none;
+  font-size: 40px;
+  text-align: center;
+}
