+++ webapp/wordle/src/components/Wordle/Wordle.css
@@ -19,9 +19,3 @@
   display: flex;
   flex-direction: column;
 }
-
-.cells {
-  height: 55px;
-  width: 55px;
-  border: none;
-}
