+++ webapp/wordle/src/components/Wordle/Wordle.js
@@ -1,24 +1,11 @@
 import React, { useState } from 'react';
 import './Wordle.css';
 import Keyboard from '../Keyboard/Keyboard.js';
+import Cells from '../Cells/Cells.js';
 
 function Wordle() {
     const rows = 5;
     const cols = 6;
-
-    const [inputValue, setInputValue] = useState("");
-    const [whichKey, setWhichKey] = useState("");
-
-    const handleKeyDown = (event) => {
-        setWhichKey(event.key);
-    }
-
-    const handleChange = (event) => {
-        if (inputValue.length >= 1 && whichKey !== "Backspace")
-            return;
-        else 
-            setInputValue(event.target.value);
-    }
 
     return (
         <>
@@ -27,7 +14,7 @@ function Wordle() {
                     <div key={rowIndex} className="row">
                         {Array.from({ length: cols }, (_, colIndex) => (
                             <div key={colIndex} className="box">
-                                <input className="cells" type="text" value={inputValue} onChange={handleChange}  onKeyDown={handleKeyDown} />
+                                <Cells />
                             </div>
                         ))}
                     </div>
